Rework primitive types; WIP

+++ bench/addictive_addition/addictive_addition.java
@@ -1,7 +1,7 @@
 class AddictiveAddition {
 
     public static void main(String[] args) {
-        int i = 0;
+        i64 i = 0;
 
         while (i < 10_000_000) {
             i++;
